% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_prefix}
\alias{add_prefix}
\title{Add a prefix to a vector of variable names}
\usage{
add_prefix(
  x = c("nav_lon", "nav_lat", "Bathy_depth", "land_mask", "dXbtm", "dXbtm_ann", "dMLD",
    "dMLD_ann", "dSST", "dSST_ann", "dSSS", "dSSS_ann", "dTbtm", "dTbtm_ann", "dSbtm",
    "dSbtm_ann", "dU", "dU_ann", "dV", "dV_ann"),
  prefix = "d",
  except = c("nav_lon", "nav_lat", "Bathy_depth", "land_mask")
)
}
\arguments{
\item{x}{character vector of variable names}

\item{prefix}{character the prefix to add}

\item{except}{character vector of names to *not* modify}
}
\value{
character vector of same length as \code{x} but with prefix added
}
\description{
Add a prefix to a vector of variable names
}
