% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_suffix}
\alias{add_suffix}
\title{Add a suffix to a vector of variable names}
\usage{
add_suffix(
  x = c("nav_lon", "nav_lat", "Bathy_depth", "land_mask", "dXbtm", "dXbtm_ann", "dMLD",
    "dMLD_ann", "dSST", "dSST_ann", "dSSS", "dSSS_ann", "dTbtm", "dTbtm_ann", "dSbtm",
    "dSbtm_ann", "dU", "dU_ann", "dV", "dV_ann"),
  suffix = "_ann",
  except = c("nav_lon", "nav_lat", "Bathy_depth", "land_mask")
)
}
\arguments{
\item{x}{character vector of variable names}

\item{suffix}{character the prefix to add}

\item{except}{character vector of names to *not* modify}
}
\value{
character vector of same length as \code{x} but with suffix added
}
\description{
Add a suffix to a vector of variable names
}
