% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{compose_filename}
\alias{compose_filename}
\title{Compose a Brickman filepath}
\usage{
compose_filename(
  scenario = c("RCP45", "RCP85", "PRESENT")[1],
  year = c(2055, 2075, NA)[1],
  path = get_path("nc")
)
}
\arguments{
\item{scenario}{character, (case insensitive) one of 'RCP45', 'RCP85', or 'PRESENT'}

\item{year}{integer or character, the year to read, ignored if scenario is 'PRESENT'}

\item{path}{charcater, the root path to the data}
}
\description{
Compose a Brickman filepath
}
\examples{
\dontrun{
read_layers()
}
}
