% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_points}
\alias{extract_points}
\title{Extract values from NCDF by point location}
\usage{
extract_points(X, vars, pts)
}
\arguments{
\item{X}{ncdf4 object}

\item{vars}{one or more variable names}

\item{pts}{sf POINT object}
}
\value{
tibble with one row per point requested and column per var requested
  Note that monthly vars will return a list column (12 elements for each point)
}
\description{
No check is performed to test if a point falls within the extent of the grid. Use this
only for curvilinear grids; for regular rectilinear grids use \code{\link[stars]{st_extract}}.
}
