% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_bb}
\alias{get_bb}
\title{Retrieve convenient bounding boxes}
\usage{
get_bb(name = c("nwa", "native")[2])
}
\arguments{
\item{name}{char the name of the bounding box, defaults to "native"}
}
\value{
\code{sf::st_sfc()} object
}
\description{
Retrieve convenient bounding boxes
}
