% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_path}
\alias{get_path}
\title{Retrieve a path for the project data}
\usage{
get_path(..., root = "/mnt/ecocast/coredata/brickman")
}
\arguments{
\item{...}{file path segements, see \code{\link[base]{file.path}}}

\item{path}{character the root directory}
}
\value{
path description for a directory or file
}
\description{
Retrieve a path for the project data
}
