% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{mask_brickman}
\alias{mask_brickman}
\title{Mask a brickman stars object}
\usage{
mask_brickman(x, mask = read_brickman_mask())
}
\arguments{
\item{x}{stars object}

\item{mask}{stars object of mask}
}
\value{
the input but mask
}
\description{
Mask a brickman stars object
}
