% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{query_layers}
\alias{query_layers}
\title{Retrieve a listing of the variables in a Brickman file}
\usage{
query_layers(
  scenario = c("RCP45", "RCP85", "PRESENT")[1],
  year = c(2055, 2075, NA)[1],
  path = get_path("nc")
)
}
\arguments{
\item{scenario}{character, (case insensitive) one of 'RCP45', 'RCP85', or 'PRESENT'}

\item{year}{integer or character, the year to read, ignored if scenario is 'PRESENT'}

\item{path}{character, the root path to the data}
}
\value{
character vector of variable names
}
\description{
Retrieve a listing of the variables in a Brickman file
}
