% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_brickman}
\alias{read_brickman}
\title{Easy Brickman ncdf data into stars, tibble or sf objects}
\usage{
read_brickman(
  scenario = c("RCP45", "RCP85", "PRESENT")[1],
  year = c(2055, 2075, NA)[1],
  vars = c("SST", "SSS"),
  interval = c("ann", "mon")[1],
  path = get_path("nc"),
  verbose = FALSE,
  add = NULL,
  form = c("stars", "tibble", "sf")[1]
)
}
\arguments{
\item{scenario}{character, (case insensitive) one of 'RCP45', 'RCP85', or 'PRESENT'}

\item{year}{integer or character, the year to read, ignored if scenario is 'PRESENT'}

\item{vars}{character, variable names (leading "d" and no trailing "_ann")}

\item{interval}{character, on of "ann" or "mon"}

\item{path}{character, the root path to the data}

\item{verbose}{logical, if TRUE output messages}

\item{add}{stars object or NULL  If not NULL then add this to the result.
Note, Bathy_depth, land_mask, nav_lon and nav_lat variables are excluded, and
only variables common to both the add and the file read are added.}

\item{form}{character, one of 'stars', 'tibble' or 'sf' to control output form}
}
\description{
Lazily fixes variable names to drop "d" prefix and "_ann" suffix which permits
user to request variables ala "SST", "SSS" etc regardless of scenario.
}
\examples{
\dontrun{
present <- read_brickman(scenario = "PRESENT")
rcp45_2075 <- read_brickman(year = 2075, form = "tibble", add = present)
}
}
