% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_brickman_mask}
\alias{read_brickman_mask}
\title{Read just the land mask}
\usage{
read_brickman_mask(land = NA, path = get_path("nc"))
}
\arguments{
\item{land}{NA (default) or "native" if NA then convert land values to NA}

\item{path}{character, the root path to the data}
}
\value{
stars object
}
\description{
Read just the land mask
}
