% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_layers}
\alias{read_layers}
\title{Read Brickman ncdf data into stars objects}
\usage{
read_layers(
  scenario = c("RCP45", "RCP85", "PRESENT")[1],
  year = c(2055, 2075, NA)[1],
  vars = c("dSST", "dSSS"),
  interval = c("ann", "mon")[1],
  path = get_path("nc"),
  verbose = FALSE
)
}
\arguments{
\item{scenario}{character, (case insensitive) one of 'RCP45', 'RCP85', or 'PRESENT'}

\item{year}{integer or character, the year to read, ignored if scenario is 'PRESENT'}

\item{vars}{character, variable names (leading "d" and no trailing "_ann")}

\item{interval}{charcater, on of "ann" or "mon"}

\item{path}{charcater, the root path to the data}

\item{verbose}{logical, if TRUE output messages}
}
\description{
Read Brickman ncdf data into stars objects
}
\examples{
\dontrun{
read_layers()
}
}
