% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{strip_prefix}
\alias{strip_prefix}
\title{Remove a prefix from a vector of variable names}
\usage{
strip_prefix(
  x = c("nav_lon", "nav_lat", "Bathy_depth", "land_mask", "dXbtm", "dXbtm_ann", "dMLD",
    "dMLD_ann", "dSST", "dSST_ann", "dSSS", "dSSS_ann", "dTbtm", "dTbtm_ann", "dSbtm",
    "dSbtm_ann", "dU", "dU_ann", "dV", "dV_ann"),
  prefix = "d"
)
}
\arguments{
\item{x}{character vector of variable names}

\item{prefix}{character, what to strip}
}
\value{
character vector of same length as \code{x} but with prefix stripped
}
\description{
Remove a prefix from a vector of variable names
}
