% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{warp_brickman}
\alias{warp_brickman}
\title{Warp a Brickman stars object}
\usage{
warp_brickman(x, crs = get_bb("nwa"))
}
\arguments{
\item{x}{stars object}

\item{crs}{st_crs object}
}
\value{
warped stars object
}
\description{
Warp a Brickman stars object
}
